Generalized add_risk_meters script
The current script requires code changes to be operational and has a lot of custom code that looks like it was built for a single use case. I generalized the script by stripping it down to just accepting a name and query string. Also added an end of script completion message.