Update get_metrics.rb
Corrected a few lines that were broken following changes to the metrics endpoint. Also corrected a few other bugs and added additional debugging. Note that I also filed an issue where the script is not working when a meter list is not provided. IMO that is okay, but the documentation says otherwise. Don't know if this has ever worked or not.